requirements.txt
fastmcp>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
websockets>=10.0
requests>=2.26.0
pydantic>=1.8.0
redis>=4.0.0  # для кеширования
pytest>=6.2.5  # для тестирования
openai>=0.27.0  # для интеграции с OpenAI
jinja2>=3.0.0  # для шаблонов промптов
jsonschema>=4.0.0  # для валидации JSON
pyyaml>=6.0  # для работы с YAML в метаданных промптов 